requirements.txt
requests~=2.31.0
requests-oauthlib~=1.3.1
redis~=4.6.0
flask~=2.3.2
beautifulsoup4~=4.12.2
openai~=0.27.8
python-dotenv~=1.0.0
markupsafe==2.1.3
tweepy~=4.14.0
oauthlib~=3.2.2
Pillow~=10.0.0
